Rename app_events.ts to appEvents.ts and update imports
