+++ public/app/angular/services/UtilSrv.ts
@@ -1,5 +1,5 @@
 import { deprecationWarning } from '@grafana/data';
-import { appEvents } from 'app/core/app_events';
+import { appEvents } from 'app/core/appEvents';
 
 import { HideModalEvent, ShowModalEvent } from '../../types/events';
