+++ public/app/core/appEvents.ts
