+++ public/app/features/alerting/unified/components/rule-editor/VizWrapper.tsx
@@ -6,7 +6,7 @@ import { FieldConfigSource, GrafanaTheme2, PanelData, ThresholdsConfig } from '@
 import { PanelRenderer } from '@grafana/runtime';
 import { GraphFieldConfig, GraphTresholdsStyleMode } from '@grafana/schema';
 import { PanelContext, PanelContextProvider, useStyles2 } from '@grafana/ui';
-import appEvents from 'app/core/app_events';
+import appEvents from 'app/core/appEvents';
 import { PanelOptions } from 'app/plugins/panel/table/models.gen';
 
 import { useVizHeight } from '../../hooks/useVizHeight';
