+++ public/app/features/datasources/settings/DataSourceSettingsPage.tsx
@@ -5,7 +5,7 @@ import { DataSourceSettings, urlUtil } from '@grafana/data';
 import { selectors } from '@grafana/e2e-selectors';
 import { Alert, Button } from '@grafana/ui';
 import { cleanUpAction } from 'app/core/actions/cleanUp';
-import appEvents from 'app/core/app_events';
+import appEvents from 'app/core/appEvents';
 import Page from 'app/core/components/Page/Page';
 import { contextSrv } from 'app/core/core';
 import { GrafanaRouteComponentProps } from 'app/core/navigation/types';
