+++ public/app/features/manage-dashboards/DashboardImportPage.tsx
@@ -19,7 +19,7 @@ import {
   VerticalGroup,
   withTheme2,
 } from '@grafana/ui';
-import appEvents from 'app/core/app_events';
+import appEvents from 'app/core/appEvents';
 import Page from 'app/core/components/Page/Page';
 import { GrafanaRouteComponentProps } from 'app/core/navigation/types';
 import { getNavModel } from 'app/core/selectors/navModel';
