+++ public/app/features/playlist/ShareModal.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 
 import { AppEvents, SelectableValue, UrlQueryMap, urlUtil } from '@grafana/data';
 import { Checkbox, ClipboardButton, Field, FieldSet, Icon, Input, Modal, RadioButtonGroup } from '@grafana/ui';
-import appEvents from 'app/core/app_events';
+import appEvents from 'app/core/appEvents';
 
 import { buildBaseUrl } from '../dashboard/components/ShareModal/utils';
