+++ public/app/features/plugins/datasource_srv.ts
@@ -16,7 +16,7 @@ import {
   getBackendSrv,
 } from '@grafana/runtime';
 import { ExpressionDatasourceRef } from '@grafana/runtime/src/utils/DataSourceWithBackend';
-import appEvents from 'app/core/app_events';
+import appEvents from 'app/core/appEvents';
 import config from 'app/core/config';
 import {
   dataSource as expressionDatasource,
